Temp restructure of analysis
